+++ packages/brick-kit/src/core/Router.ts
@@ -1,3 +1,4 @@
+import { locationsAreEqual, createPath } from "history";
 import { PluginLocation } from "@easyops/brick-types";
 import {
   loadScript,
@@ -20,20 +21,39 @@ import { getHistory } from "../history";
 import { httpErrorToString, handleHttpError } from "../handleHttpError";
 import { isUnauthenticatedError } from "../isUnauthenticatedError";
 import { brickTemplateRegistry } from "./TemplateRegistries";
+import { RecentApps } from "./interfaces";
 
 export class Router {
   private defaultCollapsed = false;
   private locationContext: LocationContext;
   private rendering = false;
   private nextLocation: PluginLocation;
+  private prevLocation: PluginLocation;
 
   constructor(private kernel: Kernel) {}
 
   async bootstrap(): Promise<void> {
     const history = getHistory();
+    this.prevLocation = history.location;
     history.listen(async (location: PluginLocation) => {
-      if (location.state && location.state.notify === false) {
-        // No rendering if notify is `false`.
+      let ignoreRendering = false;
+      const omittedLocationProps: any = {
+        hash: undefined,
+        key: undefined
+      };
+      if (
+        locationsAreEqual(
+          { ...this.prevLocation, ...omittedLocationProps },
+          { ...location, ...omittedLocationProps }
+        ) ||
+        (location.state && location.state.notify === false)
+      ) {
+        // Ignore rendering if location not changed except hash and key.
+        // Ignore rendering if notify is `false`.
+        ignoreRendering = true;
+      }
+      this.prevLocation = location;
+      if (ignoreRendering) {
         return;
       }
       if (this.rendering) {
@@ -119,7 +139,11 @@ export class Router {
       }
     }
 
-    const { mountPoints, currentApp: previousApp } = this.kernel;
+    const {
+      mountPoints,
+      currentApp: previousApp,
+      currentUrl: previouseUrl
+    } = this.kernel;
     const currentApp = storyboard ? storyboard.app : undefined;
     // Storyboard maybe re-assigned, e.g. when open launchpad.
     const appChanged =
@@ -192,11 +216,17 @@ export class Router {
 
       if (appChanged) {
         this.kernel.currentApp = currentApp;
+      }
+      this.kernel.currentUrl = createPath(location);
+      this.kernel.updateWorkspaceStack();
+
+      if (appChanged) {
         window.dispatchEvent(
-          new CustomEvent("app.change", {
+          new CustomEvent<RecentApps>("app.change", {
             detail: {
               previousApp,
-              currentApp
+              currentApp,
+              previousWorkspace: this.kernel.getPreviousWorkspace()
             }
           })
         );