+++ packages/build-config-factory/src/ensureDeps.js
@@ -1,30 +1,31 @@
 const path = require("path");
+const fs = require("fs");
 const {
   scanBricksInStoryboard,
   scanTemplatesInStoryboard
 } = require("@easyops/brick-utils");
+const yaml = require("js-yaml");
 
 module.exports = function ensureDeps() {
-  const microPackage = require(path.join(process.cwd(), "package.json"));
-
-  const microStoryboard = require(path.join(process.cwd(), "storyboard.json"));
+  const packageJson = require(path.resolve("package.json"));
+  const storyboardJson = require(path.resolve("storyboard.json"));
 
   // 排除掉原生的 html 标签，custom-element 都会带 -
-  const requiredBricks = scanBricksInStoryboard(microStoryboard).filter(item =>
+  const requiredBricks = scanBricksInStoryboard(storyboardJson).filter(item =>
     item.includes("-")
   );
 
-  const requiredTemplates = scanTemplatesInStoryboard(microStoryboard);
+  const requiredTemplates = scanTemplatesInStoryboard(storyboardJson);
 
-  const peerDependencies = Object.keys(microPackage.peerDependencies || []);
+  const peerDependencies = Object.keys(packageJson.peerDependencies || []);
 
-  const importedPackages = microStoryboard.imports || [];
+  const importedPackages = storyboardJson.imports || [];
 
   // 校验 imports 字段中 package 是否在 peerDependencies 声明
   importedPackages.forEach(pkg => {
     if (!peerDependencies.includes(pkg)) {
       throw new Error(
-        `Can't find ${pkg} module, please add it to peerDependencies of "${microPackage.name}"`
+        `Can't find ${pkg} module, please add it to peerDependencies of "${packageJson.name}"`
       );
     }
   });
@@ -63,7 +64,7 @@ module.exports = function ensureDeps() {
   requiredBricks.forEach(brick => {
     if (!importedAllBricks.has(brick)) {
       throw new Error(
-        `The custom element "${brick}" is used, but none of dependent packages defined it in "${microPackage.name}"`
+        `The custom element "${brick}" is used, but none of dependent packages defined it in "${packageJson.name}"`
       );
     }
   });
@@ -72,8 +73,26 @@ module.exports = function ensureDeps() {
   requiredTemplates.forEach(template => {
     if (!importedAllTemplates.has(template)) {
       throw new Error(
-        `The template "${template}" is used, but none of dependent packages defined it in "${microPackage.name}"`
+        `The template "${template}" is used, but none of dependent packages defined it in "${packageJson.name}"`
       );
     }
   });
+
+  // 校验 `package.conf.yaml` 中是否包含了 `*-NB` `*-NT` 的构件，它们应当声明在 `package.json` 的 `peerDependencies` 中。
+  const confPath = path.resolve("deploy-default/package.conf.yaml");
+  if (!fs.existsSync(confPath)) {
+    return;
+  }
+
+  const conf = yaml.safeLoad(fs.readFileSync(confPath, "utf8"));
+  const unexpectedDependencies = conf.dependencies
+    .map(dep => dep.name)
+    .filter(name => name.endsWith("-NB") || name.endsWith("-NT"));
+  if (unexpectedDependencies.length > 0) {
+    throw new Error(
+      `Unexpected dependencies in "package.conf.yaml": ${unexpectedDependencies.join(
+        ","
+      )}. *-NB and *-NT dependencies should always be declared in "package.json"'s "peerDependencies".`
+    );
+  }
 };