+++ packages/build-config-factory/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.6](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@1.2.5...@easyops/build-config-factory@1.2.6) (2019-12-06)
+
+### Bug Fixes
+
+- validate dependencies in package.conf.yaml ([7290d87](https://git.easyops.local/anyclouds/next-core/commits/7290d87))
+
+## [1.2.5](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@1.2.4...@easyops/build-config-factory@1.2.5) (2019-12-05)
+
+### Bug Fixes
+
+- ensure reserved directories ([af320eb](https://git.easyops.local/anyclouds/next-core/commits/af320eb))
+
 ## [1.2.4](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@1.2.3...@easyops/build-config-factory@1.2.4) (2019-11-27)
 
 **Note:** Version bump only for package @easyops/build-config-factory