+++ packages/custom-antd-styles/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.1](https://git.easyops.local/anyclouds/next-core/compare/@easyops/custom-antd-styles@0.8.0...@easyops/custom-antd-styles@0.8.1) (2019-11-28)
+
+### Bug Fixes
+
+- update menu styles ([58f1558](https://git.easyops.local/anyclouds/next-core/commits/58f1558))
+
+# [0.8.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/custom-antd-styles@0.7.1...@easyops/custom-antd-styles@0.8.0) (2019-11-28)
+
+### Features
+
+- modify checkbox indeterminate status style ([7a58152](https://git.easyops.local/anyclouds/next-core/commits/7a58152))
+
+## [0.7.1](https://git.easyops.local/anyclouds/next-core/compare/@easyops/custom-antd-styles@0.7.0...@easyops/custom-antd-styles@0.7.1) (2019-11-28)
+
+### Bug Fixes
+
+- update card padding ([46edba0](https://git.easyops.local/anyclouds/next-core/commits/46edba0))
+
 # [0.7.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/custom-antd-styles@0.6.0...@easyops/custom-antd-styles@0.7.0) (2019-11-27)
 
 ### Features