+++ packages/rollup-config-factory/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.3](https://git.easyops.local/anyclouds/next-core/compare/@easyops/rollup-config-factory@1.0.2...@easyops/rollup-config-factory@1.0.3) (2019-11-28)
+
+### Bug Fixes
+
+- use content hash instead of default for postcss modules ([792fc60](https://git.easyops.local/anyclouds/next-core/commits/792fc60))
+
 ## [1.0.2](https://git.easyops.local/anyclouds/next-core/compare/@easyops/rollup-config-factory@1.0.1...@easyops/rollup-config-factory@1.0.2) (2019-11-27)
 
 **Note:** Version bump only for package @easyops/rollup-config-factory